+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.1] - 2025-10-17
+
+### Fixed
+
+- Fixed Y-axis tick values to only consider future prices when not starting at midnight
+
+### Changed
+
+- Improved code organization and documentation in renderer.py
+
 ## [0.1.0] - 2025-10-17
 
 ### Added